software update checking added
git-svn-id: http://svn.assembla.com/svn/symphonie/trunk@97 57038eb5-193d-0410-9cfd-c93977fb03f4

+++ pjproject-0.5.10.4/pjmedia/src/pjmedia-codec/speex/jitter.c
@@ -181,7 +181,7 @@ void jitter_buffer_put(JitterBuffer *jitter, const JitterBufferPacket *packet)
    
    /* Copy packet in buffer */
    jitter->buf[i]=(char*)speex_alloc(packet->len);
-   for (j=0;j<packet->len;j++)
+   for (j=0;j<(int)packet->len;j++)
       jitter->buf[i][j]=packet->data[j];
    jitter->timestamp[i]=packet->timestamp;
    jitter->span[i]=packet->span;
@@ -378,7 +378,7 @@ int jitter_buffer_get(JitterBuffer *jitter, JitterBufferPacket *packet, spx_uint
       /* Check for potential overflow */
       packet->len = jitter->len[i];
       /* Copy packet */
-      for (j=0;j<packet->len;j++)
+      for (j=0;j<(int)packet->len;j++)
          packet->data[j] = jitter->buf[i][j];
       /* Remove packet */
       speex_free(jitter->buf[i]);
